ridge-link-sled/pyproject.toml
[project]
name = "ridge-link-sled"
version = "1.0.0"
description = "Lightweight agent for racing simulators"
dependencies = [
    "requests",
    "psutil",
    "opencv-python",
    "pillow",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
